requirements.txt
pandas
requests
rembg
pillow
onnxruntime
tqdm